+++ src/app/actions/color.tsx
@@ -7,7 +7,10 @@ import colorValidator from "@/validator/server/color";
 import { revalidateTag } from "next/cache";
 
 export async function getColors() {
-  const res = await fetch(`${apiUrl}/color`, {
+  if (!apiUrl) {
+    return null;
+  }
+  const res = await fetch(`${apiUrl}/api/color`, {
     next: { tags: ["colors"] },
   });
   const colors = await res.json();
@@ -16,7 +19,10 @@ export async function getColors() {
 
 export async function deleteColor(id: string) {
   "use server";
-  const res = await fetch(`${apiUrl}/color/${id}`, {
+  if (!apiUrl) {
+    return null;
+  }
+  const res = await fetch(`${apiUrl}/api/color/${id}`, {
     method: "DELETE",
   });
 
@@ -26,6 +32,9 @@ export async function deleteColor(id: string) {
 
 export async function createColors(prev: any, formData: FormData) {
   "use server";
+  if (!apiUrl) {
+    return null;
+  }
   const data = {
     code: formData.get("code"),
     name: formData.get("name"),
@@ -65,6 +74,9 @@ export async function createColors(prev: any, formData: FormData) {
 
 export async function editColors(prev: any, formData: FormData) {
   "use server";
+  if (!apiUrl) {
+    return null;
+  }
   await connectToDB();
   const id = formData.get("id");
